+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * perf: `ObjectToObjectTransformer` optimization
 * perf: move `TypeResolver` outside of the hot path
 * perf(`ObjectCache`): use sentinel value instead of exception
+* feat: extra target values
 
 ## 1.10.0